% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{assert_dates_in_time_line}
\alias{assert_dates_in_time_line}
\title{Assert that the dates are available within the time_line}
\usage{
assert_dates_in_time_line(dates, tl)
}
\arguments{
\item{dates}{Dates to check}

\item{tl}{A [time_line] object to check against}
}
\description{
Assert that the dates are available within the time_line
}
