% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{assert_poll_data_in_time_line}
\alias{assert_poll_data_in_time_line}
\title{Assert that the poll_data object dates are available
within the time_line}
\usage{
assert_poll_data_in_time_line(pd, tl)
}
\arguments{
\item{pd}{[polls_data] to check if part of [time_line]}

\item{tl}{[time_line] to check against}
}
\description{
Assert that the poll_data object dates are available
within the time_line
}
