% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_slurm.R
\name{bash_script_remove_files}
\alias{bash_script_remove_files}
\title{Generate a script to remove all files in a folder}
\usage{
bash_script_remove_files(directory, ...)
}
\arguments{
\item{directory}{the directory with files to remove}

\item{...}{further arguments to [dir]}
}
\description{
Generate a script to remove all files in a folder
}
