% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{calibration_plot}
\alias{calibration_plot}
\alias{calibration_qqplot}
\title{Calibration plots for simulated data with knon state}
\usage{
calibration_plot(x, true_ls, alpha = 0.05, ...)

calibration_qqplot(x, true_ls, ...)
}
\arguments{
\item{x}{a list of [poll_of_polls] object with the same model}

\item{true_ls}{a vector with the true state}

\item{alpha}{the alpha level for the CI}

\item{...}{further arguments to [geom_histogram]}
}
\description{
Calibration plots for simulated data with knon state
}
