% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{complete_poll_info}
\alias{complete_poll_info}
\alias{complete_poll_data}
\alias{complete_poll_y}
\alias{complete_polls}
\title{Return indicator on polls without missing values}
\usage{
complete_poll_info(x)

complete_poll_data(x)

complete_poll_y(x)

complete_polls(x)
}
\arguments{
\item{x}{a [polls_data] object.}
}
\value{
a boolean vector
}
\description{
Return indicator on polls without missing values
}
\details{
[complete_poll_info()] indicate polls with complete poll_info.

[complete_poll_data()] and [complete_poll_y()] indicate polls with complete poll data (or y)

[complete_polls()] indicate polls with complete poll data and info
}
