% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_arguments.R
\name{compute_g_scale_default}
\alias{compute_g_scale_default}
\title{Compute the default value of g_scale}
\usage{
compute_g_scale_default(x)
}
\arguments{
\item{x}{a list with g and next_known_state_index}
}
\description{
Compute the default value of g_scale
}
