% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_arguments.R
\name{compute_gamma_parameters}
\alias{compute_gamma_parameters}
\title{Compute gamma parameters with same expectation and variances
as a half-normal with parameter sigma}
\usage{
compute_gamma_parameters(sigma, n = 1e+06)
}
\arguments{
\item{sigma}{the sigma parameter in a half-formal}

\item{n}{how many samples to approximate the half-normal}
}
\value{
a list with parameters a and b
}
\description{
Compute gamma parameters with same expectation and variances
as a half-normal with parameter sigma
}
