% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_errors.R
\name{compute_prediction_error}
\alias{compute_prediction_error}
\title{Compute prediction error for a poll_of_polls model}
\usage{
compute_prediction_error(x, predict_state, type = "elpd")
}
\arguments{
\item{x}{a [poll_of_polls] model}

\item{predict_state}{the true value to predict}

\item{type}{error measurement to compute}
}
\description{
Compute prediction error for a poll_of_polls model
}
