% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{dates_in_time_line}
\alias{dates_in_time_line}
\title{What dates are available within the time_line}
\usage{
dates_in_time_line(dates, tl)
}
\arguments{
\item{dates}{Dates to check}

\item{tl}{A [time_line] object to check against}
}
\description{
What dates are available within the time_line
}
