% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_extract_min_eigen_value_Omega}
\alias{diagnose_extract_min_eigen_value_Omega}
\title{Extract minimal eigen value per iteration}
\usage{
diagnose_extract_min_eigen_value_Omega(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}
}
\description{
Extract minimal eigen value per iteration
}
