% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{divergent_transitions_percentiles}
\alias{divergent_transitions_percentiles}
\title{Extract the percentiles of the divergent transistions}
\usage{
divergent_transitions_percentiles(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}
}
\value{
a matrix of the percentiles of the divergent transitions
}
\description{
Extract the percentiles of the divergent transistions
}
