% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{elpd_known_state}
\alias{elpd_known_state}
\alias{elpd_known_state.pop_model8f1}
\alias{elpd_known_state.pop_model8f2}
\alias{rmse_known_state}
\alias{percentile_known_state}
\title{Compute the elpd, percentiles and rmse for a true [known_state]}
\usage{
elpd_known_state(x, known_state)

\method{elpd_known_state}{pop_model8f1}(x, known_state)

\method{elpd_known_state}{pop_model8f2}(x, known_state)

rmse_known_state(x, known_state)

percentile_known_state(x, known_state)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{known_state}{a [known_state] object}
}
\description{
Compute the elpd, percentiles and rmse for a true [known_state]
}
\details{
Incorrect draws has been removed
}
