% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{elpd_nlfo}
\alias{elpd_nlfo}
\alias{elpd_nlfo.poll_of_polls}
\title{Compute the naive leave future out elpd
compared with a true latent state.}
\usage{
elpd_nlfo(x, true_ls, model, t_0, type = "full", ...)

\method{elpd_nlfo}{poll_of_polls}(x, true_ls, model, t_0 = 2, type = "full", ...)
}
\arguments{
\item{x}{a poll_of_polls object.}

\item{true_ls}{the true latent state.}

\item{model}{Model to compute elpd_nlfo for}

\item{t_0}{Starting timepoint to compute from.}

\item{type}{the type of elpd_nlfo to compute [full] (total elpd),
and [series] (a series of elpd_t)}

\item{...}{further arguments (currently not used).}
}
\description{
Compute the naive leave future out elpd
compared with a true latent state.
}
