% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_slurm.R
\name{existing_parties}
\alias{existing_parties}
\title{Return parties existing in a given model range}
\usage{
existing_parties(y, ltr, mtr)
}
\arguments{
\item{y}{a character vector of party names}

\item{ltr}{the [time_range] of the latent series}

\item{mtr}{the [time_range] of the model}
}
\description{
Return parties existing in a given model range
}
