% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract}
\alias{extract}
\alias{extract.poll_of_polls}
\title{Extract parameters from the stan object}
\usage{
extract(object, ...)

\method{extract}{poll_of_polls}(object, ...)
}
\arguments{
\item{object}{an object to extract parameter draws from}

\item{...}{arguments supplied to rstan::extract.}
}
\description{
Extract parameters from the stan object
}
