% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_diagnostics.R
\name{extract_diagnostics_as_df}
\alias{extract_diagnostics_as_df}
\title{Extract diagnostics as data frames}
\usage{
extract_diagnostics_as_df(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}
}
\description{
Extract diagnostics as data frames
}
