% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_known_state}
\alias{geom_known_state}
\title{Visualize a known state in a poll_of_polls ggplot}
\usage{
geom_known_state(x, y, ...)
}
\arguments{
\item{x}{a known_states data.frame}

\item{y}{the specific variable to visualize}

\item{...}{further arguments to geom_line and geom_point}
}
\description{
Visualize a known state in a poll_of_polls ggplot
}
