% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_latent_state}
\alias{geom_latent_state}
\title{Add a [latent_state] geom to a ggplot}
\usage{
geom_latent_state(
  x,
  median = TRUE,
  intervals = c(0.9, 0.75, 0.5),
  latent_state_colour = "darkgrey",
  ...
)
}
\arguments{
\item{x}{a [latent_state] object}

\item{median}{should the median be plotted?}

\item{intervals}{what intervals should be visualized}

\item{latent_state_colour}{the color of the latent state}

\item{...}{further arguments to [geom_ribbon] and [geom_line].}
}
\description{
Add a [latent_state] geom to a ggplot
}
