% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_stan_polls_data}
\alias{geom_stan_polls_data}
\title{Visualize a stan_data object}
\usage{
geom_stan_polls_data(x, ...)
}
\arguments{
\item{x}{a [stan_polls_data] object}

\item{...}{further arguments supplied to [geom_polls_time_weights()].}
}
\description{
Visualize a stan_data object
}
\seealso{
stan_polls_data
}
