% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{german_elections}
\alias{german_elections}
\title{German election data}
\format{
A data frame with 4 elections, 2005-2017
}
\source{
\url{http://www.wikipedia.com}
}
\usage{
german_elections
}
\description{
German election data
}
\keyword{datasets}
