% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{german_polls}
\alias{german_polls}
\alias{german_polls_curated}
\title{German polls data}
\format{
A data frame with 4609 polls and 6 parties:

An object of class \code{data.frame} with 4606 rows and 15 columns.
}
\source{
\url{http://www.wahlrecht.de/umfragen/}
}
\usage{
german_polls

german_polls_curated
}
\description{
German polls data
}
\keyword{datasets}
