% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_adaptation_info}
\alias{get_adaptation_info}
\title{Get adaptation info}
\usage{
get_adaptation_info(x, as.list = FALSE, ...)
}
\arguments{
\item{x}{a [poll_of_polls]}

\item{as.list}{return an R list. Defaults to [FALSE].}

\item{...}{further arguemt past to [rstan::get_adaptation_info]}
}
\description{
Extract the adaptation info (step size and diagonal invers mass matrix)
from a [poll_of_polls] object
}
