% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_arguments.R
\name{get_first_poll_as_simplex}
\alias{get_first_poll_as_simplex}
\title{The first poll is the poll with the earliest starting date}
\usage{
get_first_poll_as_simplex(x)
}
\arguments{
\item{x}{a stan_data object}
}
\description{
The first poll is the poll with the earliest starting date
}
