% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_state.R
\name{get_latent_state_for_dates}
\alias{get_latent_state_for_dates}
\title{Get the latent state for specific dates}
\usage{
get_latent_state_for_dates(x, date)
}
\arguments{
\item{x}{a [poll_of_polls] or [latent_state] object}

\item{date}{a [date] vector}
}
\description{
Get the latent state for specific dates
}
