% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{get_model_diagnostics}
\alias{get_model_diagnostics}
\title{Get the model diagnostics used in a [poll_of_polls] object}
\usage{
get_model_diagnostics(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}
}
\value{
a list with diagnostic values
}
\description{
Get the model diagnostics used in a [poll_of_polls] object
}
\details{
no_divergent_transitions: The number of draws with divergent transitions
no_max_treedepth: The number of draws where the max treedepth was reached
no_low_bfmi_chains: The number of chains with low E-BFMI
no_Rhat_above_1_1: The number of parameters with an Rhat above 1.1
no_Rhat_is_NA: The number of parameters without an Rhat value
mean_no_leapfrog_steps: The mean number of leapfrog steps taken per iteration/draw
mean_chain_step_size: The mean stepsize over the chains
mean_chain_inv_mass_matrix_min_max: The mean max and min of the invers mass matrix over the chains
mean_chain_warmup_time: the mean warmup time in seconds
mean_chain_sampling_time: the mean warmup time in seconds
}
