% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{get_model_model_arguments}
\alias{get_model_model_arguments}
\title{Get the model arguments used in a [poll_of_polls] object}
\usage{
get_model_model_arguments(x, all = TRUE)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{all}{Should all model arguments be returned, or just the one that were set?}
}
\value{
a list with model arguments
}
\description{
Get the model arguments used in a [poll_of_polls] object
}
