% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_slurm.R
\name{get_model_periods}
\alias{get_model_periods}
\title{Get the periods to use in modeling}
\usage{
get_model_periods(cfg, known_state)
}
\arguments{
\item{cfg}{a cfg}

\item{known_state}{a known state_object}
}
\description{
Get the periods to use in modeling
}
