% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{get_ndraws}
\alias{get_ndraws}
\alias{ndraws}
\title{Get the number of draws/samples in a [poll_of_polls] object}
\usage{
get_ndraws(x)

ndraws(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}
}
\description{
Get the number of draws/samples in a [poll_of_polls] object
}
