% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{get_polls_next_known_state_index}
\alias{get_polls_next_known_state_index}
\alias{get_time_line_next_known_state_index}
\alias{get_dates_next_known_state_index}
\title{Get the next known state index}
\usage{
get_polls_next_known_state_index(x, known_state, type = "collection_midpoint")

get_time_line_next_known_state_index(time_line, known_state)

get_dates_next_known_state_index(dates, known_state)
}
\arguments{
\item{x}{a [polls_data] object}

\item{known_state}{a [known_state] object}

\item{type}{how is the next state index defined.
Default is based on [collection_period] dates.}

\item{time_line}{time line to get next index for}

\item{dates}{Dates to get next index for}
}
\description{
Get the next known state index
}
\details{
The function returns the index for the next known state.
Polls without a next unknown state, gets index [nrow(known_state) + 1].
}
