% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{get_polls_time_points_between_known_states}
\alias{get_polls_time_points_between_known_states}
\alias{get_time_line_time_points_between_known_states}
\alias{get_dates_time_points_between_known_states}
\title{Get the number of time points between known state}
\usage{
get_polls_time_points_between_known_states(
  x,
  known_state,
  tl,
  type = "collection_midpoint"
)

get_time_line_time_points_between_known_states(tl, known_state)

get_dates_time_points_between_known_states(dates, known_state, tl)
}
\arguments{
\item{x}{a [polls_data] object}

\item{known_state}{a [known_state] object}

\item{tl}{a [time_line] object to get the number of time points between known state}

\item{type}{how is the next state index defined.
Default is based on [collection_period] dates.}

\item{dates}{Dates to get the number of time points between known states}
}
\description{
Get the number of time points between known state
}
\details{
The function returns the number of time points between
the known states.
}
