% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{get_polls_time_points_since_last_known_state}
\alias{get_polls_time_points_since_last_known_state}
\alias{get_time_line_time_points_since_last_known_state}
\alias{get_dates_time_points_since_last_known_state}
\title{Get the number of time points since the last known state}
\usage{
get_polls_time_points_since_last_known_state(
  x,
  known_state,
  tl,
  type = "collection_midpoint"
)

get_time_line_time_points_since_last_known_state(tl, known_state)

get_dates_time_points_since_last_known_state(dates, known_state, tl)
}
\arguments{
\item{x}{a [polls_data] object}

\item{known_state}{a [known_state] object}

\item{tl}{a [time_line] object to get the number of time points since last known state}

\item{type}{how is the next state index defined.
Default is based on [collection_period] dates.}

\item{dates}{Dates to get the number of time points since the last known state}
}
\description{
Get the number of time points since the last known state
}
\details{
The function returns the number of time points since the last known state.
}
