% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_get_functions.R
\name{get_sampler_params}
\alias{get_sampler_params}
\alias{get_sampler_params.poll_of_polls}
\title{Get Sampler Diagnostics}
\usage{
get_sampler_params(object, ...)

\method{get_sampler_params}{poll_of_polls}(object, ...)
}
\arguments{
\item{object}{a [poll_of_polls] object to extract stan samples from}

\item{...}{arguments further supplied to [rstan::get_sampler_params()]}
}
\description{
Get Sampler Diagnostics
}
\seealso{
rstan::get_sampler_params
}
