% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_get_functions.R
\name{get_stancode}
\alias{get_stancode}
\alias{get_stancode.poll_of_polls}
\title{Get Stan Code from a pop object}
\usage{
get_stancode(object, ...)

\method{get_stancode}{poll_of_polls}(object, ...)
}
\arguments{
\item{object}{a [poll_of_polls] object to extract stan code from}

\item{...}{arguments further supplied to [rstan::get_stancode()]}
}
\description{
Get Stan Code from a pop object
}
\seealso{
rstan::get_stancode
}
