% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{get_time_points}
\alias{get_time_points}
\alias{get_time_points.time_line}
\alias{get_time_points.poll_of_polls}
\title{Get the time points for specific dates}
\usage{
get_time_points(x, dates)

\method{get_time_points}{time_line}(x, dates)

\method{get_time_points}{poll_of_polls}(x, dates)
}
\arguments{
\item{x}{a [time_line] or [poll_of_polls]}

\item{dates}{a [date] vector}
}
\description{
Get the time points for specific dates
}
