% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{house_map}
\alias{house_map}
\title{Map between houses and house names}
\format{
A data.frame with house, house name and country.:
}
\usage{
house_map
}
\description{
Map between houses and house names
}
\keyword{datasets}
