% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_state.R
\name{latent_state}
\alias{latent_state}
\alias{latent_state.poll_of_polls}
\alias{latent_state.stanfit}
\title{Create a latent state object}
\usage{
latent_state(x, time_line = NULL, ...)

\method{latent_state}{poll_of_polls}(x, time_line = NULL, ...)

\method{latent_state}{stanfit}(x, time_line, dimnames, ...)
}
\arguments{
\item{x}{an object to convert to a [latent_state]}

\item{time_line}{a [time_line] object}

\item{...}{further arguments to methods}

\item{dimnames}{The dimnames attribute for the latent state array}
}
\description{
Create a latent state object
}
