% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_state.R
\name{latent_state_abind}
\alias{latent_state_abind}
\title{Bind latent_state objects}
\usage{
latent_state_abind(x, y, along, ...)
}
\arguments{
\item{x}{a [latent_state] object}

\item{y}{a [latent_state] object}

\item{along}{the dimension to bind the latent states. See [abind::abind] for details.}

\item{...}{further arguments to [abind::abind]}
}
\value{
a [latent_state] object
}
\description{
Bind latent_state objects
}
