% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_state.R
\name{latent_state_apply}
\alias{latent_state_apply}
\title{Apply latent_state objects}
\usage{
latent_state_apply(x, MARGIN, FUN, ...)
}
\arguments{
\item{x}{a [latent_state] object to subet}

\item{MARGIN}{margins to apply over}

\item{FUN}{function to apply}

\item{...}{further arguments to base::apply}
}
\value{
a [latent_state] object
}
\description{
Apply latent_state objects
}
