% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{latent_state_rmse}
\alias{latent_state_rmse}
\alias{latent_state_rmse.poll_of_polls}
\alias{latent_state_rmse.latent_state}
\title{Compute the RMSE for the latent state
compared with a true latent state.}
\usage{
latent_state_rmse(x, true_ls, type = "series", ...)

\method{latent_state_rmse}{poll_of_polls}(x, true_ls, type = "series", ...)

\method{latent_state_rmse}{latent_state}(x, true_ls, type = "series", ...)
}
\arguments{
\item{x}{a [latent_state] object.}

\item{true_ls}{the true latent state}

\item{type}{the type of RMSE to compute [full] (a total RMSE),
and [series] (a series with RMSE over time)}

\item{...}{further arguments (currently not used).}
}
\description{
Compute the RMSE for the latent state
compared with a true latent state.
}
