% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_state.R
\name{latent_state_subset}
\alias{latent_state_subset}
\title{Subset latent_state objects}
\usage{
latent_state_subset(x, iterations = NULL, t = NULL, categories = NULL)
}
\arguments{
\item{x}{a [latent_state] object to subet}

\item{iterations}{subset iterations}

\item{t}{subset time points (t)}

\item{categories}{subset categories}
}
\value{
a [latent_state] object
}
\description{
Subset latent_state objects
}
