% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_slurm.R
\name{load_models}
\alias{load_models}
\title{Load models from model output directory}
\usage{
load_models(dir)
}
\arguments{
\item{dir}{directury with model files}
}
\description{
Load models from model output directory
}
