% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{make_job_name}
\alias{make_job_name}
\alias{unmake_job_name}
\title{Make a name froma a grid expand matrix}
\usage{
make_job_name(x, i, sep = "=", collapse = "|")

unmake_job_name(x, sep = "=", collapse = "\\\\|")
}
\arguments{
\item{x}{a [data.frame] from [grid.expand()]}

\item{i}{the row to make name for}

\item{sep}{char to separate argument names and values}

\item{collapse}{char to separate arguments}
}
\description{
Function to make a job name from a [grid.expand()]
}
