% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_arguments.R
\name{model_config}
\alias{model_config}
\alias{model_arguments}
\title{Set default values and assert arguments for a model configuration}
\usage{
model_config(model, x = NULL, stan_data = NULL)

model_arguments(model)
}
\arguments{
\item{model}{the model for which to set the model arguments}

\item{x}{a list with model arguments (the model config) set by the user}

\item{stan_data}{a list with the stan data used for the model}
}
\description{
Set default values and assert arguments for a model configuration
}
\details{
With model arguments, the arguments that are set by the
user for each poll_of_polls stan model. This differs between
models. See [model_arguments()] for the model arguments
used by each model.
}
