% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_dates_data_frame}
\alias{model_dates_data_frame}
\title{Return a data.frame with dates from a list of pop objects}
\usage{
model_dates_data_frame(x)
}
\arguments{
\item{x}{a list of pop objects}
}
\description{
Return a data.frame with dates from a list of pop objects
}
