% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{on_local_branch}
\alias{on_local_branch}
\alias{on_local_test_branch}
\alias{on_github_actions}
\alias{on_github_actions_branch}
\alias{on_github_actions_test_branch}
\alias{get_working_directory_path}
\title{Check if the tests are run in a specific git branch
or on a certain CI.}
\usage{
on_local_branch(name)

on_local_test_branch()

on_github_actions()

on_github_actions_branch(name)

on_github_actions_test_branch()

get_working_directory_path()
}
\arguments{
\item{name}{The name of a git branch to check}
}
\description{
Check if the tests are run in a specific git branch
or on a certain CI.
}
\details{
If not in git repo, FALSE is returned.
}
