% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{parameter_names}
\alias{parameter_names}
\alias{parameter_block_names}
\alias{parameters_counts}
\alias{get_parameter_names}
\title{Extract the parameter names and parameter counts from a poll_of_polls object}
\usage{
parameter_names(x)

parameter_block_names(x)

parameters_counts(x)

get_parameter_names(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}
}
\description{
Extract the parameter names and parameter counts from a poll_of_polls object
}
