% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{parameter_names_from_unconstrained_index}
\alias{parameter_names_from_unconstrained_index}
\alias{parameter_block_names_from_unconstrained_index}
\alias{parameter_unique_names_from_unconstrained_index}
\title{Get parameter names from unconstrained space}
\usage{
parameter_names_from_unconstrained_index(stanfit, up_idx)

parameter_block_names_from_unconstrained_index(stanfit, up_idx)

parameter_unique_names_from_unconstrained_index(stanfit, up_idx)
}
\arguments{
\item{stanfit}{a [stanfit] object to extract parameter names from}

\item{up_idx}{the index of parameters at the constrained space to extract names for}
}
\value{
a [data.frame] with one row per parameter affected by the unconstrained
}
\description{
The function returns the name and indecies of all parameters on the constrained
space that are affected by the unconstrained parameter.
}
