% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{parameter_summary}
\alias{parameter_summary}
\title{Extract a parameter summary of interest}
\usage{
parameter_summary(x, parameter_name, summary_type)
}
\arguments{
\item{x}{an object to extract from}

\item{parameter_name}{to select (rownames of summary(stanfit))}

\item{summary_type}{to select (colnames of summary(stanfit))}
}
\description{
Extract a parameter summary of interest
}
