% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party_color.R
\name{party_color}
\alias{party_color}
\title{Get the party color for a given party abbrivation}
\usage{
party_color(x)
}
\arguments{
\item{x}{a party abbrivation}
}
\description{
Get the party color for a given party abbrivation
}
