% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{pd_test}
\alias{pd_test}
\title{A subset of swe_polls stored as a [polls_data] object}
\format{
a [polls_data] object with 1007 polls.
}
\usage{
pd_test
}
\description{
A subset of swe_polls stored as a [polls_data] object
}
\details{
All polls from [swe_polls] without missing data
during 2010-01-01 to 2019-12-31.
See data-raw folder for exact construction.
}
\keyword{datasets}
