% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.polls_data}
\alias{plot.polls_data}
\alias{geom_publish_date}
\alias{geom_collection_period}
\title{Plot a [polls_data] object}
\usage{
\method{plot}{polls_data}(x, y = NULL, publish_date = TRUE, collection_period = TRUE, ...)

geom_publish_date(x, y, ...)

geom_collection_period(x, y, ...)
}
\arguments{
\item{x}{a [polls_data] object}

\item{y}{a variable name in the [polls_data] object.}

\item{publish_date}{visualize publication dates.}

\item{collection_period}{visualize collection period.}

\item{...}{further arguments to [geom_collection_period]}
}
\description{
Plot a [polls_data] object
}
