% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_kappa_vs_x}
\alias{plot_kappa_vs_x}
\title{Plot a x vs kappa plot}
\usage{
plot_kappa_vs_x(pop, y, t)
}
\arguments{
\item{pop}{a [poll_of_polls] object with industry bias using kappa.}

\item{y}{a variable name in the [poll_of_polls] object.}

\item{t}{time line time point to plot}
}
\description{
Plot a x vs kappa plot
}
