% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_parameters_bayesplot}
\alias{plot_parameters_bayesplot}
\title{Plot posterior distributions of parameters}
\usage{
plot_parameters_bayesplot(
  x,
  bayeplot_FUN,
  params,
  params_plot_name = NULL,
  ...
)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{bayeplot_FUN}{a plotting function from the basyeplot package}

\item{params}{parameters to plot (see \code{parameter_names()} too see all parameters in model)}

\item{params_plot_name}{names to use for parameters in plot}

\item{...}{further arguments sent to bayesplot::mcmc_area}
}
\description{
Plot posterior distributions of parameters
}
