% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls_data.R
\name{poll_id}
\alias{poll_id}
\alias{poll_ids}
\alias{poll_id<-}
\alias{poll_ids<-}
\alias{start_date}
\alias{start_dates}
\alias{start_date<-}
\alias{start_dates<-}
\alias{end_date}
\alias{end_dates}
\alias{end_date<-}
\alias{end_dates<-}
\alias{publish_date}
\alias{publish_dates}
\alias{publish_date<-}
\alias{publish_dates<-}
\alias{n}
\alias{collection_midpoint_dates}
\alias{houses}
\alias{n<-}
\alias{y}
\alias{y<-}
\title{Get and set elements of a [data_polls] object}
\usage{
poll_id(x)

poll_ids(x)

poll_id(x) <- value

poll_ids(x) <- value

start_date(x)

start_dates(x)

start_date(x) <- value

start_dates(x) <- value

end_date(x)

end_dates(x)

end_date(x) <- value

end_dates(x) <- value

publish_date(x)

publish_dates(x)

publish_date(x) <- value

publish_dates(x) <- value

n(x)

collection_midpoint_dates(x)

houses(x)

n(x) <- value

y(x)

y(x) <- value
}
\arguments{
\item{x}{a [data_polls] object.}

\item{value}{a value to set.}
}
\description{
Get and set elements of a [data_polls] object
}
