% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls_data.R
\name{polls_data}
\alias{polls_data}
\alias{as.polls_data}
\alias{[.polls_data}
\title{Create a [polls_data] object}
\usage{
polls_data(y, house, publish_date, start_date, end_date, n, poll_id = NULL)

as.polls_data(x, ...)

\method{[}{polls_data}(x, i, keep_poll_id = TRUE, ...)
}
\arguments{
\item{y}{a named [data.frame] with variables of interest to estimate over time.}

\item{house}{a factor variable indicating polling house (of same length as rows in [y])}

\item{publish_date}{the date of when the poll was published}

\item{start_date}{the start date of the data collection period of each poll}

\item{end_date}{the end date of the data collection period of each poll}

\item{n}{the (integer) sample size of each poll}

\item{poll_id}{the id of each poll. Deafult is [row.names] of [y].}

\item{x}{an object  to convert to a [polls_data] object.}

\item{...}{further arguments to methods.}

\item{i}{row in [polls_data] to select.}

\item{keep_poll_id}{Should the [poll_id]}
}
\description{
Create a [polls_data] object
}
\examples{
data(swedish_polls_curated)
pd <- polls_data(y = swedish_polls_curated[,3:11],
                 house = swedish_polls_curated$Company,
                 publish_date = swedish_polls_curated$PublDate,
                 start_date = swedish_polls_curated$collectPeriodFrom,
                 end_date = swedish_polls_curated$collectPeriodTo,
                 n = swedish_polls_curated$n)

}
