% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls_time_weights.R
\name{polls_time_reweight}
\alias{polls_time_reweight}
\title{Reweights the weights  based on function}
\usage{
polls_time_reweight(x, poll_ids, FUN, ...)
}
\arguments{
\item{x}{a [polls_time_weights] to reweight.}

\item{poll_ids}{a character vector with poll_ids to reweight.}

\item{FUN}{function used to reweight}

\item{...}{further arguments to [FUN].}
}
\description{
Reweights the weights  based on function
}
\details{
The function [FUN] is used to compute the weights
for integer values 0,...,T_i, where T_i is the length
of the period for poll i. The values from poll is then
normalized to sum to 1.
}
