% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/polls_time_weights.R
\name{geom_polls_time_weights}
\alias{geom_polls_time_weights}
\alias{polls_time_weights}
\alias{polls_time_weights.polls_data}
\alias{polls_time_weights.stan_polls_data}
\alias{polls_time_weights<-}
\alias{polls_time_weights<-.polls_data}
\title{Compute a time period weight tibble}
\usage{
geom_polls_time_weights(x, y, ...)

polls_time_weights(x, ...)

\method{polls_time_weights}{polls_data}(x, week_start = 1, ...)

\method{polls_time_weights}{stan_polls_data}(x, ...)

polls_time_weights(x) <- value

\method{polls_time_weights}{polls_data}(x) <- value
}
\arguments{
\item{x}{a [polls_data] object.}

\item{y}{a vector to plot for the [polls_time_weights] object.}

\item{...}{further arguments to [geom_point]}

\item{week_start}{day on which week starts following ISO
conventions - 1 means Monday, 7 means Sunday (default).}

\item{value}{a [polls_time_weights] object.}
}
\value{
a [polls_time_weights] object (tibble)
}
\description{
Constructs a tibble with the weight of each day of a poll.
The tibble can then be aggregated to the right level time
scale.
The first day of each year, month and week is used to set a
date for each time point.
}
