% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recompile_stanfit.R
\name{recompile_stanfit}
\alias{recompile_stanfit}
\title{Recompile a poll_of_polls object}
\usage{
recompile_stanfit(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}
}
\value{
a recompiled [stanfit] object based on the model and data of the [poll_of_polls] object
}
\description{
The function takes a [poll_of_polls] object and recomplie the code an the data
}
