% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_polls_data_model.R
\name{remove_known_states}
\alias{remove_known_states}
\title{function for removing known states outside timeline}
\usage{
remove_known_states(
  psd,
  known_state,
  model_time_range = NULL,
  time_scale = "week"
)
}
\arguments{
\item{psd}{[polls_data]}

\item{known_state}{data.frame containg known result needs date (date of the known data)}

\item{model_time_range}{time range of where model is active}

\item{time_scale}{what scale is the data at}
}
\description{
function for removing known states outside timeline
}
