% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{reweight_and_resample}
\alias{reweight_and_resample}
\title{Reweight and resample a polls_data object}
\usage{
reweight_and_resample(x, true_ls, time_scale, weight = "none")
}
\arguments{
\item{x}{a [polls_data] object.}

\item{true_ls}{a vector with the true latent state}

\item{time_scale}{The [time_scale] used for the latent state.}

\item{weight}{The way the collection period is weighted. See code for details.}
}
\description{
The functions reweights and resamples a data_polls object.
This is used to simulated polls with different [time_weights].
}
