% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{Rhat_is_above_m_by_parameter_block}
\alias{Rhat_is_above_m_by_parameter_block}
\alias{Rhat_is_above_m}
\alias{Rhat_is_NA_by_parameter_block}
\alias{Rhat_is_NA}
\title{Extract more detailed diagnostic information}
\usage{
Rhat_is_above_m_by_parameter_block(x, m = 1.1)

Rhat_is_above_m(x, m = 1.1)

Rhat_is_NA_by_parameter_block(x)

Rhat_is_NA(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{m}{a limit to extract Rhat values above}
}
\description{
Extract more detailed diagnostic information
}
