% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{run_simulation_experiments}
\alias{run_simulation_experiments}
\title{Run Simulation experiment with different settings}
\usage{
run_simulation_experiments(
  jobs,
  pd,
  values = NULL,
  cache_folder = NULL,
  figure_folder = NULL,
  verbose = TRUE,
  sim_start_date = "2010-01-04"
)
}
\arguments{
\item{jobs}{a data frame with one job per row and columns as argument}

\item{pd}{The [polls_data] object to use for simulated data.}

\item{values}{Values of interest to compute as "rmse", "elpd_nlfo" and "sigma_x|mean"
where the last indicates that we want the mean summary for parameter "sigma_x"}

\item{cache_folder}{folder with cached [poll_of_polls] objects. NULL means no cache.}

\item{figure_folder}{folder with cached [poll_of_polls] objects. NULL means no figures.}

\item{verbose}{The simulation experiment is verbose.}

\item{sim_start_date}{Date to start data in simulations.}
}
\description{
Run Simulation experiment with different settings
}
