% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polls.R
\name{sample_polls_y}
\alias{sample_polls_y}
\title{Sample a dependent variable based on the polls}
\usage{
sample_polls_y(x, true_ls, tl, week_start = 1L)
}
\arguments{
\item{x}{a [polls_data] object}

\item{true_ls}{a numeric vector of the length as the [time_line]}

\item{tl}{a  [time_line] object.}

\item{week_start}{the integer day starting the week (1 = Monday, 7 = Sunday).}
}
\description{
Sample a dependent variable based on the polls
}
\details{
The [y] is sampled  based on the [time_weights] and
the [true_ls].
}
