% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_helpers.R
\name{set_collection_period}
\alias{set_collection_period}
\title{Set collection period}
\usage{
set_collection_period(x, type = "interval")
}
\arguments{
\item{x}{a [polls_data] object}

\item{type}{type of collection period to set.}
}
\description{
Set the collection period to test different approaches

* [mid_collection]: Set all dates to mid of collection period
* [publish_date]: Set all dates to publish date
* [interval]: Use the collection period interval (Default)
}
