% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_time_ranges.R
\name{setup_latent_time_ranges}
\alias{setup_latent_time_ranges}
\title{Construct a list of [time_range]s}
\usage{
setup_latent_time_ranges(x, y, mtr, check_dates = TRUE)
}
\arguments{
\item{x}{a list with a character vector with either "from" or "to"}

\item{y}{all categories as a chacrater vector}

\item{mtr}{a model [time_range] that will be used as default}

\item{check_dates}{assert that the dates of [x] is in the [mtr] interval}
}
\description{
Construct a list of [time_range]s
}
