% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polls.R
\name{simulate_polls}
\alias{simulate_polls}
\title{Simulate Polls}
\usage{
simulate_polls(
  x,
  pd,
  npolls,
  time_scale = "week",
  start_date = as.Date("2010-01-04")
)
}
\arguments{
\item{x}{a latent series to simulate polls for. Either a numeric vector or a data.frame.}

\item{pd}{a [polls_data] object to sample poll
periods and sample sizes from.}

\item{npolls}{The total number of polls to simulate}

\item{time_scale}{The time_scale used, [day], [week], [month] or [year].}

\item{start_date}{The starting date for the polls.}
}
\description{
Simulate [npolls] polls for a given latent series [x],
where the time slace is defined by [time_scale] and starting
from [start_date].
Polling periods are randomly sampled from a a [polls_data] object.
}
