% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_slurm.R
\name{slurm_model_runs}
\alias{slurm_model_runs}
\alias{generate_sbatch_from_path}
\alias{bash_script_sbatch_files}
\alias{triton_model_runs}
\title{Generate sbatch scripts to run all files in a folder}
\usage{
slurm_model_runs(directory)

generate_sbatch_from_path(directory)

bash_script_sbatch_files(directory)

triton_model_runs(directory = "triton/model_runs")
}
\arguments{
\item{directory}{a directory with files to submit with batch}
}
\description{
Generate sbatch scripts to run all files in a folder
}
