% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{spanish_elections}
\alias{spanish_elections}
\title{Spanish election data}
\format{
A data frame with 6 elections, 2008-2019
}
\source{
\url{http://www.wikipedia.com}
}
\usage{
spanish_elections
}
\description{
Spanish election data
}
\keyword{datasets}
