% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{swedish_polls_curated}
\alias{swedish_polls_curated}
\alias{spanish_polls}
\alias{spanish_polls_curated}
\title{Spanish polls data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1660 rows and 18 columns.

A data frame with 1852 polls and 7 parties:

An object of class \code{data.frame} with 1345 rows and 16 columns.
}
\source{
\url{http://www.wikipedia.com}
}
\usage{
swedish_polls_curated

spanish_polls

spanish_polls_curated
}
\description{
Spanish polls data
}
\keyword{datasets}
