% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_polls_data_model.R
\name{stan_data_add_missing}
\alias{stan_data_add_missing}
\title{Adds missing data through NA to stan data}
\usage{
stan_data_add_missing(stan_data)
}
\arguments{
\item{stan_data}{contatning $y which can include missing in the form of na}
}
\description{
Adds missing data through NA to stan data
}
