% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{stan_non_state_parameters}
\alias{stan_non_state_parameters}
\title{Parameters that are not state parameters, by model}
\usage{
stan_non_state_parameters(model)
}
\arguments{
\item{model}{a model name to get non-state parameters from.}
}
\description{
Parameters that are not state parameters, by model
}
\details{
The function returns the parameter names of the parameters
that are not state parameters and hence can be visualized.
}
\keyword{internal}
