% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{stan_parameters_to_store}
\alias{stan_parameters_to_store}
\title{Parameters we want to store, by model}
\usage{
stan_parameters_to_store(model)
}
\arguments{
\item{model}{a model name to get the parameters to store from.}
}
\description{
Parameters we want to store, by model
}
\details{
The function returns the parameter names of the parameters that are needed
for analysis or in computing predictive distributions.
Hence these parameters will be the only ones that will be stored
when running Stan.
}
\keyword{internal}
