% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls_data.R
\name{standard_error}
\alias{standard_error}
\title{Compute standard errors for poll estimates}
\usage{
standard_error(x, y, type = "binomial")
}
\arguments{
\item{x}{a [polls_data] object}

\item{y}{variable to compute standrad error for}

\item{type}{type of standard error estimator}
}
\description{
Compute standard errors for poll estimates
}
