% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls_data.R
\name{subset.polls_data}
\alias{subset.polls_data}
\title{Subset a [polls_data] object.}
\usage{
\method{subset}{polls_data}(x, subset, select, drop = FALSE, ...)
}
\arguments{
\item{x}{a [polls_data]}

\item{subset, select, drop, ...}{see [subset()]}
}
\description{
Subset a [polls_data] object.
}
