% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_latent_state_dates}
\alias{subset_latent_state_dates}
\alias{subset_time_line_dates}
\title{Subset a latent state based on a time range}
\usage{
subset_latent_state_dates(x, from = NULL, to = NULL, ...)

subset_time_line_dates(x, from = NULL, to = NULL, ...)
}
\arguments{
\item{x}{a [latent_state] object}

\item{from}{subset from [date]}

\item{to}{subset to [date]}

\item{...}{Currently not in use.}
}
\description{
Subset a latent state based on a time range
}
