% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_publish_dates}
\alias{subset_publish_dates}
\alias{subset_dates}
\title{Subset a [polls_data] object based on a specific [time_range]}
\usage{
subset_publish_dates(x, from = NULL, to = NULL, ...)

subset_dates(x, from = NULL, to = NULL, ...)
}
\arguments{
\item{x}{a [polls_data] object}

\item{from}{a the earliest date to include polls for}

\item{to}{a the latest date to include polls for}

\item{...}{further arguments}
}
\description{
Subset a [polls_data] object.
* [subset_publish_dates()] subset the data with the polls
  that has  a  publication date within the interval.
* [subset_dates()] subset the data with the polls
  that has all non-[NA] dates in the interval.
}
