% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls_time_weights.R
\name{summarize_polls_time_weights}
\alias{summarize_polls_time_weights}
\title{Summarize a polls weight object by a time line}
\usage{
summarize_polls_time_weights(ptw, tl)
}
\arguments{
\item{ptw}{a [polls_time_weights] object.}

\item{tl}{a [time_line] object.}
}
\value{
a [polls_time_weights] object.
}
\description{
Summarize a polls weight object by a time line
}
