% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{swedish_elections}
\alias{swedish_elections}
\title{Swedish election data}
\format{
A data frame with 4 elections, 2006-2018
}
\source{
\url{http://www.wikipedia.com}
}
\usage{
swedish_elections
}
\description{
Swedish election data
}
\keyword{datasets}
