% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{swedish_party_time_ranges}
\alias{swedish_party_time_ranges}
\alias{german_party_time_ranges}
\alias{spanish_party_time_ranges}
\title{Party time ranges}
\format{
A named list (parties) with starting and ending dates.

An object of class \code{list} of length 1.

An object of class \code{list} of length 2.
}
\usage{
swedish_party_time_ranges

german_party_time_ranges

spanish_party_time_ranges
}
\description{
Party time ranges
}
\details{
Time ranges for political parties
}
\keyword{datasets}
