% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{swedish_polls}
\alias{swedish_polls}
\title{Swedish polls data}
\format{
A data frame with 1565 polls and 9 parties:
}
\source{
\url{http://www.github.com/MansMeg/SwedishPolls}
}
\usage{
swedish_polls
}
\description{
Swedish polls data
}
\keyword{datasets}
