% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_prediction_errors_results.R
\name{table_latent_mean_coverage}
\alias{table_latent_mean_coverage}
\title{Compute latent mean coverage}
\usage{
table_latent_mean_coverage(x, standard_errors = 1)
}
\arguments{
\item{x}{a poll of polls object (x$pops)}

\item{standard_errors}{the number of standard errors to use.}
}
\description{
Computes coverage of polls with respect to the latent mean
}
