% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_diagnostics.R
\name{table_models_parameter_diagnostics}
\alias{table_models_parameter_diagnostics}
\alias{table_models_model_diagnostics}
\title{Compute parameter and sampler diagnostics tables}
\usage{
table_models_parameter_diagnostics(x)

table_models_model_diagnostics(x)
}
\arguments{
\item{x}{a list with models used in report 3}
}
\description{
Compute parameter and sampler diagnostics tables
}
