% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_prediction_errors_results.R
\name{table_predict_known_state_error}
\alias{table_predict_known_state_error}
\alias{predict_known_state_error_table}
\title{Compute error tables}
\usage{
table_predict_known_state_error(x, known_states, type)

predict_known_state_error_table(x, known_states, type)
}
\arguments{
\item{x}{a list of poll of polls models (x$pops)}

\item{known_states}{a dataframe with known states for the parties}

\item{type}{type of errors to compute (rmse, quantile, elpd)}
}
\description{
Compute prediction tables for a given set of models and the known
true values
}
