% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_pop_models_env}
\alias{test_pop_models_env}
\alias{set_test_pop_models_env}
\title{Get and set TEST_POP_MODELS env variable}
\usage{
test_pop_models_env()

set_test_pop_models_env(x)
}
\arguments{
\item{x}{value to set it to}
}
\description{
Get and set TEST_POP_MODELS env variable
}
