% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{time_line_add_slow_scale}
\alias{time_line_add_slow_scale}
\title{Create a time line with slower scales}
\usage{
time_line_add_slow_scale(tl, slow_scales)
}
\arguments{
\item{tl}{a [time_line] object}

\item{slow_scales}{a vector of [Date]s that indicate breaks (right-inclusive) for a slower moving time scale.
Example: If only 2010-01-15 is used, all dates up to and including 2010-01-15, will have s=1,
         Dates after 2010-01-15 will have s=2.}
}
\description{
Create a time line with slower scales
}
\keyword{internal}
