% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{time_line_expand}
\alias{time_line_expand}
\title{Expand a time_line object to a new time range
but keep the same time point index.}
\usage{
time_line_expand(tl, time_range)
}
\arguments{
\item{tl}{a time_line to expand}

\item{time_range}{a new time range to expand to}
}
\description{
Expand a time_line object to a new time range
but keep the same time point index.
}
