% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_range.R
\name{time_range}
\alias{time_range}
\alias{time_range<-}
\alias{time_range_polls}
\title{Time range}
\usage{
time_range(x, ...)

time_range(x) <- value

time_range_polls(x, ...)
}
\arguments{
\item{x}{an object to compute the time range for.}

\item{...}{further arguments supplied to method.}

\item{value}{a Date vector of length 2 with names [from] and [to].}
}
\description{
Time range
}
\details{
Compute the time range for an object with [time_range] or
compute the time range of the polls in the object with
[time_range_polls].
}
