% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_range.R
\name{time_ranges_are_disjoint}
\alias{time_ranges_are_disjoint}
\title{Check if two time ranges are disjoint (not overlapping)}
\usage{
time_ranges_are_disjoint(tr1, tr2)
}
\arguments{
\item{tr1}{first [time_range]}

\item{tr2}{second [time_range]}
}
\description{
Check if two time ranges are disjoint (not overlapping)
}
