% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_bash_generation.R
\name{uppmax_bash_args}
\alias{uppmax_bash_args}
\title{Extract arguments from bash base}
\usage{
uppmax_bash_args(return_name = TRUE)
}
\arguments{
\item{return_name}{return the argument name}
}
\description{
Extract arguments from bash base
}
