% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_latent_mean_csv}
\alias{write_latent_mean_csv}
\alias{write_latent_draw_csv}
\alias{write_ada_json}
\alias{write_ada_time_series_json}
\alias{write_ada_latent_posterior_json}
\title{Write poll of polls models}
\usage{
write_latent_mean_csv(x, file)

write_latent_draw_csv(x, file, draw_idx)

write_ada_json(
  x,
  output_dir = getwd(),
  file_name = NULL,
  percentiles = c(0.05, 0.125, 0.5, 0.875, 0.95)
)

write_ada_time_series_json(
  x,
  output_dir = getwd(),
  file_name = NULL,
  percentiles = c(0.05, 0.125, 0.5, 0.875, 0.95),
  aggregated_series = list()
)

write_ada_latent_posterior_json(
  x,
  output_dir = getwd(),
  file_name = NULL,
  last_no_t = NULL
)
}
\arguments{
\item{x}{a [poll_of_polls] object to write}

\item{file}{the file path}

\item{draw_idx}{draw index to write.}

\item{output_dir}{output directory}

\item{file_name}{the output file name.}

\item{percentiles}{percentiles to output}

\item{aggregated_series}{a list with combinations of aggregations to make (e.g. S, V and MP). Default is no aggregations.}

\item{last_no_t}{last number of time points to include in output}
}
\description{
Write poll of polls models
}
