% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_bash_generation.R
\name{write_uppmax_bash}
\alias{write_uppmax_bash}
\title{Write a UPPMAX bash file from an argument list}
\usage{
write_uppmax_bash(args, directory, model_config, stan_arguments, stan_control)
}
\arguments{
\item{args}{a list with relevant arguments to put in [bash_code_base()]}

\item{directory}{path to output files}

\item{model_config}{a list with model arguments}

\item{stan_arguments}{a list with stan arguments}

\item{stan_control}{a list with stan control arguments}
}
\description{
Write a UPPMAX bash file from an argument list
}
