% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_predict.R
\name{x_predict}
\alias{x_predict}
\alias{x_predict.pop_model6b}
\alias{x_predict.pop_model6c}
\title{Compute the posterior predictive distribution of polls}
\usage{
x_predict(x, y = NULL, t = NULL, ...)

\method{x_predict}{pop_model6b}(x, y = NULL, t = NULL, ...)

\method{x_predict}{pop_model6c}(x, y = NULL, t = NULL, ...)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{y}{a vector of categories to compute the predicted distribution for}

\item{t}{a vector of time points to compute the predictive for for the latent state.
If [NULL], the whole latent state is returned.
If [t] exist in latent state, the latent state is returned.}

\item{...}{Currently not in use.}
}
\description{
Compute the posterior predictive distribution of polls
}
