% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_data.R
\docType{data}
\name{x_test}
\alias{x_test}
\title{Four true latent states used in experiments}
\format{
A list of length four with numeric vectors of length 100.
}
\usage{
x_test
}
\description{
Four true latent states used in experiments
}
\details{
See data-raw folder for exact construction.
}
\keyword{datasets}
