% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_predict.R
\name{y_predict}
\alias{y_predict}
\alias{poll_predictive_distribution}
\alias{y_predict.pop_model6b}
\alias{y_predict.pop_model6c}
\title{Compute the posterior predictive distribution of polls}
\usage{
y_predict(x, y = NULL, poll_ids = NULL, ...)

poll_predictive_distribution(x, y = NULL, poll_ids = NULL, ...)

\method{y_predict}{pop_model6b}(x, y = NULL, poll_ids = NULL, ...)

\method{y_predict}{pop_model6c}(x, y = NULL, poll_ids = NULL, ...)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{y}{a vector of categories to compute the predicted distribution for}

\item{poll_ids}{a vector of poll ids to compute the predictive for. If [NULL] all are computed.}

\item{...}{Currently not used.}
}
\description{
Compute the posterior predictive distribution of polls
}
